requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

# HTML Parsing
beautifulsoup4==4.12.3
lxml==5.2.1

# AI (Gemini)
google-generativeai==0.5.4

# Env & Utils
python-dotenv==1.0.1

# Data Validation
pydantic==2.6.4

# Optional (recommended)
httpx==0.27.0

openai